Refactor main and create startup class
